+++ app/src/main/java/com/squirtles/musicroad/pick/components/DetailPickTopAppBar.kt
@@ -17,8 +17,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import com.squirtles.musicroad.R
-import com.squirtles.musicroad.map.components.CreatedByOtherUserText
-import com.squirtles.musicroad.map.components.CreatedBySelfText
+import com.squirtles.musicroad.common.CreatedByOtherUserText
+import com.squirtles.musicroad.common.CreatedBySelfText
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
