+++ app/src/main/res/values/strings.xml
@@ -16,7 +16,7 @@
     <string name="map_album_image_description">앨범 이미지</string>
     <string name="map_info_window_pick_user">님의 픽</string>
     <string name="map_info_window_favorite_count_icon_description">픽을 담은 개수</string>
-    <string name="map_pick_created_by_self">내가 등록한 픽</string>
+    <string name="pick_created_by_self">내가 등록한 픽</string>
 
     <!-- Pick -->
     <string name="pick_app_bar_title">픽 등록</string>
